+++ CHANGELOG.md
@@ -6,6 +6,9 @@ If the path contains something like:
 ```
 In this case the `dynamicElement` will be replaced, if it exists in the ScenarioContext.
 
+Support for adding static key/value pairs to the context:
+- [Set a static value to the context](README.md#set-a-static-value-to-the-context)
+- [Set multiple static values to the context](REAMDME.md#set-multiple-static-values-to-the-context)
 
 # Release 1.14.0